requirements.txt
aiohttp
pandas
beautifulsoup4
python-docx
PyPDF2